# MINOR - Review query performance