--- a/ingestion/src/metadata/ingestion/ometa/ometa_api.py
+++ b/ingestion/src/metadata/ingestion/ometa/ometa_api.py
@@ -405,8 +405,12 @@ class OpenMetadata(
                 try:
                     entities.append(entity(**elmt))
                 except Exception as exc:
-                    logger.error(f"Error creating entity [{entity.__name__}]. Failed with exception {exc}")
-                    logger.debug(f"Can't create [{entity.__name__}] from [{elmt}]. Skipping.")
+                    logger.error(
+                        f"Error creating entity [{entity.__name__}]. Failed with exception {exc}"
+                    )
+                    logger.debug(
+                        f"Can't create [{entity.__name__}] from [{elmt}]. Skipping."
+                    )
                     continue
         else:
             entities = [entity(**elmt) for elmt in resp["data"]]