--- a/openmetadata-service/src/main/java/org/openmetadata/service/jdbi3/CollectionDAO.java
+++ b/openmetadata-service/src/main/java/org/openmetadata/service/jdbi3/CollectionDAO.java
@@ -522,7 +522,7 @@ public interface CollectionDAO {
       }
 
       String sqlCondition = String.format("%s AND er.toId is NULL", condition);
-      return listCount(getTableName(), sqlCondition);
+      return listCount(getTableName(), getNameHashColumn(), sqlCondition);
     }
 
     @SqlQuery(
@@ -563,7 +563,17 @@ public interface CollectionDAO {
         @Bind("limit") int limit,
         @Bind("after") String after);
 
-    @SqlQuery(
+    @ConnectionAwareSqlQuery(
+        value =
+            "SELECT count(<nameHashColumn>) FROM <table> ce "
+                + "LEFT JOIN ("
+                + "  SELECT toId FROM entity_relationship "
+                + "  WHERE fromEntity = 'container' AND toEntity = 'container' AND relation = 0 "
+                + ") er "
+                + "on ce.id = er.toId "
+                + "<sqlCondition>",
+        connectionType = MYSQL)
+    @ConnectionAwareSqlQuery(
         value =
             "SELECT count(*) FROM <table> ce "
                 + "LEFT JOIN ("
@@ -571,8 +581,12 @@ public interface CollectionDAO {
                 + "  WHERE fromEntity = 'container' AND toEntity = 'container' AND relation = 0 "
                 + ") er "
                 + "on ce.id = er.toId "
-                + "<sqlCondition>")
-    int listCount(@Define("table") String table, @Define("sqlCondition") String mysqlCond);
+                + "<sqlCondition>",
+        connectionType = POSTGRES)
+    int listCount(
+        @Define("table") String table,
+        @Define("nameHashColumn") String nameHashColumn,
+        @Define("sqlCondition") String mysqlCond);
   }
 
   interface SearchServiceDAO extends EntityDAO<SearchService> {
@@ -1975,7 +1989,7 @@ public interface CollectionDAO {
       }
 
       String condition = filter.getCondition(getTableName());
-      return listCount(getTableName(), condition, condition);
+      return listCount(getTableName(), getNameHashColumn(), condition, condition);
     }
 
     @Override
@@ -2217,7 +2231,7 @@ public interface CollectionDAO {
 
       mySqlCondition = String.format("%s %s", mySqlCondition, filter.getCondition("tag"));
       postgresCondition = String.format("%s %s", postgresCondition, filter.getCondition("tag"));
-      return listCount(getTableName(), mySqlCondition, postgresCondition);
+      return listCount(getTableName(), getNameHashColumn(), mySqlCondition, postgresCondition);
     }
 
     @Override
@@ -2645,7 +2659,7 @@ public interface CollectionDAO {
                 "%s AND ((json#>'{isJoinable}')::boolean)  = %s ", postgresCondition, isJoinable);
       }
 
-      return listCount(getTableName(), mySqlCondition, postgresCondition);
+      return listCount(getTableName(), getNameHashColumn(), mySqlCondition, postgresCondition);
     }
 
     @Override
@@ -3386,7 +3400,8 @@ public interface CollectionDAO {
         String psqlStr = String.format("AND supported_data_types @> '`%s`' ", supportedDataType);
         psqlCondition.append(psqlStr.replace('`', '"'));
       }
-      return listCount(getTableName(), mysqlCondition.toString(), psqlCondition.toString());
+      return listCount(
+          getTableName(), getNameHashColumn(), mysqlCondition.toString(), psqlCondition.toString());
     }
 
     @ConnectionAwareSqlQuery(
@@ -3428,13 +3443,14 @@ public interface CollectionDAO {
         @Bind("after") String after);
 
     @ConnectionAwareSqlQuery(
-        value = "SELECT count(*) FROM <table> <mysqlCond>",
+        value = "SELECT count(<nameHashColumn>) FROM <table> <mysqlCond>",
         connectionType = MYSQL)
     @ConnectionAwareSqlQuery(
         value = "SELECT count(*) FROM <table> <psqlCond>",
         connectionType = POSTGRES)
     int listCount(
         @Define("table") String table,
+        @Define("nameHashColumn") String nameHashColumn,
         @Define("mysqlCond") String mysqlCond,
         @Define("psqlCond") String psqlCond);
   }
@@ -3474,7 +3490,7 @@ public interface CollectionDAO {
         postgresCondition =
             String.format("%s %s", postgresCondition, filter.getCondition(getTableName()));
       }
-      return listCount(
+      return listCountDistinct(
           getTableName(),
           mySqlCondition,
           postgresCondition,
@@ -3864,61 +3880,61 @@ public interface CollectionDAO {
   interface SystemDAO {
     @ConnectionAwareSqlQuery(
         value =
-            "SELECT (SELECT COUNT(*) FROM table_entity <cond>) as tableCount, "
-                + "(SELECT COUNT(*) FROM topic_entity <cond>) as topicCount, "
-                + "(SELECT COUNT(*) FROM dashboard_entity <cond>) as dashboardCount, "
-                + "(SELECT COUNT(*) FROM pipeline_entity <cond>) as pipelineCount, "
-                + "(SELECT COUNT(*) FROM ml_model_entity <cond>) as mlmodelCount, "
-                + "(SELECT COUNT(*) FROM storage_container_entity <cond>) as storageContainerCount, "
-                + "(SELECT COUNT(*) FROM search_index_entity <cond>) as searchIndexCount, "
-                + "(SELECT COUNT(*) FROM glossary_entity <cond>) as glossaryCount, "
-                + "(SELECT COUNT(*) FROM glossary_term_entity <cond>) as glossaryTermCount, "
-                + "(SELECT (SELECT COUNT(*) FROM metadata_service_entity <cond>) + "
-                + "(SELECT COUNT(*) FROM dbservice_entity <cond>)+"
-                + "(SELECT COUNT(*) FROM messaging_service_entity <cond>)+ "
-                + "(SELECT COUNT(*) FROM dashboard_service_entity <cond>)+ "
-                + "(SELECT COUNT(*) FROM pipeline_service_entity <cond>)+ "
-                + "(SELECT COUNT(*) FROM mlmodel_service_entity <cond>)+ "
-                + "(SELECT COUNT(*) FROM search_service_entity <cond>)+ "
-                + "(SELECT COUNT(*) FROM storage_service_entity <cond>)) as servicesCount, "
-                + "(SELECT COUNT(*) FROM user_entity <cond> AND (JSON_EXTRACT(json, '$.isBot') IS NULL OR JSON_EXTRACT(json, '$.isBot') = FALSE)) as userCount, "
-                + "(SELECT COUNT(*) FROM team_entity <cond>) as teamCount, "
-                + "(SELECT COUNT(*) FROM test_suite <cond>) as testSuiteCount",
+            "SELECT (SELECT COUNT(fqnHash) FROM table_entity <cond>) as tableCount, "
+                + "(SELECT COUNT(fqnHash) FROM topic_entity <cond>) as topicCount, "
+                + "(SELECT COUNT(fqnHash) FROM dashboard_entity <cond>) as dashboardCount, "
+                + "(SELECT COUNT(fqnHash) FROM pipeline_entity <cond>) as pipelineCount, "
+                + "(SELECT COUNT(fqnHash) FROM ml_model_entity <cond>) as mlmodelCount, "
+                + "(SELECT COUNT(fqnHash) FROM storage_container_entity <cond>) as storageContainerCount, "
+                + "(SELECT COUNT(fqnHash) FROM search_index_entity <cond>) as searchIndexCount, "
+                + "(SELECT COUNT(nameHash) FROM glossary_entity <cond>) as glossaryCount, "
+                + "(SELECT COUNT(fqnHash) FROM glossary_term_entity <cond>) as glossaryTermCount, "
+                + "(SELECT (SELECT COUNT(nameHash) FROM metadata_service_entity <cond>) + "
+                + "(SELECT COUNT(nameHash) FROM dbservice_entity <cond>)+"
+                + "(SELECT COUNT(nameHash) FROM messaging_service_entity <cond>)+ "
+                + "(SELECT COUNT(nameHash) FROM dashboard_service_entity <cond>)+ "
+                + "(SELECT COUNT(nameHash) FROM pipeline_service_entity <cond>)+ "
+                + "(SELECT COUNT(nameHash) FROM mlmodel_service_entity <cond>)+ "
+                + "(SELECT COUNT(nameHash) FROM search_service_entity <cond>)+ "
+                + "(SELECT COUNT(nameHash) FROM storage_service_entity <cond>)) as servicesCount, "
+                + "(SELECT COUNT(nameHash) FROM user_entity <cond> AND (JSON_EXTRACT(json, '$.isBot') IS NULL OR JSON_EXTRACT(json, '$.isBot') = FALSE)) as userCount, "
+                + "(SELECT COUNT(nameHash) FROM team_entity <cond>) as teamCount, "
+                + "(SELECT COUNT(fqnHash) FROM test_suite <cond>) as testSuiteCount",
         connectionType = MYSQL)
     @ConnectionAwareSqlQuery(
         value =
-            "SELECT (SELECT COUNT(*) FROM table_entity <cond>) as tableCount, "
-                + "(SELECT COUNT(*) FROM topic_entity <cond>) as topicCount, "
-                + "(SELECT COUNT(*) FROM dashboard_entity <cond>) as dashboardCount, "
-                + "(SELECT COUNT(*) FROM pipeline_entity <cond>) as pipelineCount, "
-                + "(SELECT COUNT(*) FROM ml_model_entity <cond>) as mlmodelCount, "
-                + "(SELECT COUNT(*) FROM storage_container_entity <cond>) as storageContainerCount, "
-                + "(SELECT COUNT(*) FROM search_index_entity <cond>) as searchIndexCount, "
-                + "(SELECT COUNT(*) FROM glossary_entity <cond>) as glossaryCount, "
-                + "(SELECT COUNT(*) FROM glossary_term_entity <cond>) as glossaryTermCount, "
-                + "(SELECT (SELECT COUNT(*) FROM metadata_service_entity <cond>) + "
-                + "(SELECT COUNT(*) FROM dbservice_entity <cond>)+ "
-                + "(SELECT COUNT(*) FROM messaging_service_entity <cond>)+ "
-                + "(SELECT COUNT(*) FROM dashboard_service_entity <cond>)+ "
-                + "(SELECT COUNT(*) FROM pipeline_service_entity <cond>)+ "
-                + "(SELECT COUNT(*) FROM mlmodel_service_entity <cond>)+ "
-                + "(SELECT COUNT(*) FROM search_service_entity <cond>)+ "
-                + "(SELECT COUNT(*) FROM storage_service_entity <cond>)) as servicesCount, "
-                + "(SELECT COUNT(*) FROM user_entity <cond> AND (json#>'{isBot}' IS NULL OR ((json#>'{isBot}')::boolean) = FALSE)) as userCount, "
-                + "(SELECT COUNT(*) FROM team_entity <cond>) as teamCount, "
-                + "(SELECT COUNT(*) FROM test_suite <cond>) as testSuiteCount",
+            "SELECT (SELECT COUNT(fqnHash) FROM table_entity <cond>) as tableCount, "
+                + "(SELECT COUNT(fqnHash) FROM topic_entity <cond>) as topicCount, "
+                + "(SELECT COUNT(fqnHash) FROM dashboard_entity <cond>) as dashboardCount, "
+                + "(SELECT COUNT(fqnHash) FROM pipeline_entity <cond>) as pipelineCount, "
+                + "(SELECT COUNT(fqnHash) FROM ml_model_entity <cond>) as mlmodelCount, "
+                + "(SELECT COUNT(fqnHash) FROM storage_container_entity <cond>) as storageContainerCount, "
+                + "(SELECT COUNT(fqnHash) FROM search_index_entity <cond>) as searchIndexCount, "
+                + "(SELECT COUNT(nameHash) FROM glossary_entity <cond>) as glossaryCount, "
+                + "(SELECT COUNT(fqnHash) FROM glossary_term_entity <cond>) as glossaryTermCount, "
+                + "(SELECT (SELECT COUNT(nameHash) FROM metadata_service_entity <cond>) + "
+                + "(SELECT COUNT(nameHash) FROM dbservice_entity <cond>)+"
+                + "(SELECT COUNT(nameHash) FROM messaging_service_entity <cond>)+ "
+                + "(SELECT COUNT(nameHash) FROM dashboard_service_entity <cond>)+ "
+                + "(SELECT COUNT(nameHash) FROM pipeline_service_entity <cond>)+ "
+                + "(SELECT COUNT(nameHash) FROM mlmodel_service_entity <cond>)+ "
+                + "(SELECT COUNT(nameHash) FROM search_service_entity <cond>)+ "
+                + "(SELECT COUNT(nameHash) FROM storage_service_entity <cond>)) as servicesCount, "
+                + "(SELECT COUNT(nameHash) FROM user_entity <cond> AND (json#>'{isBot}' IS NULL OR ((json#>'{isBot}')::boolean) = FALSE)) as userCount, "
+                + "(SELECT COUNT(nameHash) FROM team_entity <cond>) as teamCount, "
+                + "(SELECT COUNT(fqnHash) FROM test_suite <cond>) as testSuiteCount",
         connectionType = POSTGRES)
     @RegisterRowMapper(EntitiesCountRowMapper.class)
     EntitiesCount getAggregatedEntitiesCount(@Define("cond") String cond) throws StatementException;
 
     @SqlQuery(
-        "SELECT (SELECT COUNT(*) FROM database_entity <cond>) as databaseServiceCount, "
-            + "(SELECT COUNT(*) FROM messaging_service_entity <cond>) as messagingServiceCount, "
-            + "(SELECT COUNT(*) FROM dashboard_service_entity <cond>) as dashboardServiceCount, "
-            + "(SELECT COUNT(*) FROM pipeline_service_entity <cond>) as pipelineServiceCount, "
-            + "(SELECT COUNT(*) FROM mlmodel_service_entity <cond>) as mlModelServiceCount, "
-            + "(SELECT COUNT(*) FROM storage_service_entity <cond>) as storageServiceCount, "
-            + "(SELECT COUNT(*) FROM search_service_entity <cond>) as searchServiceCount")
+        "SELECT (SELECT COUNT(nameHash) FROM dbservice_entity <cond>) as databaseServiceCount, "
+            + "(SELECT COUNT(nameHash) FROM messaging_service_entity <cond>) as messagingServiceCount, "
+            + "(SELECT COUNT(nameHash) FROM dashboard_service_entity <cond>) as dashboardServiceCount, "
+            + "(SELECT COUNT(nameHash) FROM pipeline_service_entity <cond>) as pipelineServiceCount, "
+            + "(SELECT COUNT(nameHash) FROM mlmodel_service_entity <cond>) as mlModelServiceCount, "
+            + "(SELECT COUNT(nameHash) FROM storage_service_entity <cond>) as storageServiceCount, "
+            + "(SELECT COUNT(nameHash) FROM search_service_entity <cond>) as searchServiceCount")
     @RegisterRowMapper(ServicesCountRowMapper.class)
     ServicesCount getAggregatedServicesCount(@Define("cond") String cond) throws StatementException;
 
@@ -4290,7 +4306,8 @@ public interface CollectionDAO {
         psqlCondition.append(String.format(" AND entityType='%s' ", entityType));
       }
 
-      return listCount(getTableName(), mysqlCondition.toString(), psqlCondition.toString());
+      return listCount(
+          getTableName(), getNameHashColumn(), mysqlCondition.toString(), psqlCondition.toString());
     }
 
     @ConnectionAwareSqlQuery(
@@ -4330,17 +4347,6 @@ public interface CollectionDAO {
         @Define("psqlCond") String psqlCond,
         @Bind("limit") int limit,
         @Bind("after") String after);
-
-    @ConnectionAwareSqlQuery(
-        value = "SELECT count(*) FROM <table> <mysqlCond>",
-        connectionType = MYSQL)
-    @ConnectionAwareSqlQuery(
-        value = "SELECT count(*) FROM <table> <psqlCond>",
-        connectionType = POSTGRES)
-    int listCount(
-        @Define("table") String table,
-        @Define("mysqlCond") String mysqlCond,
-        @Define("psqlCond") String psqlCond);
   }
 
   interface SuggestionDAO {