… openmetadata-service/src/main/java/org/openmetadata/service/migration/context/CommonMigrationOps.java
… openmetadata-service/src/main/java/org/openmetadata/service/util/OpenMetadataOperations.java