<!--
Thank you for your contribution!
Unless your change is trivial, please create an issue to discuss the change before creating a PR.
-->

### Describe your changes:

The goal of this PR is to see if we can speed up backend processes with small changes on indexes / queries.

Some findings:

### 1. `COUNT WHERE deleted = FALSE`

<img width="1349" alt="image" src="https://github.com/open-metadata/OpenMetadata/assets/35870520/ba398ef8-ce7b-4163-b644-4ca1eadc2051">

In sandbox-beta (240k tables) this query was the heaviest one.

Since we rely in most of our API calls on `Include.DELETED / NOT DELETED` filter, we're adding an index on `fqnHash & deleted`. Analyzing the query after the index is applied makes it way faster.
```
-- WITH INDEX
-> Aggregate: count(0)  (cost=72102 rows=1) (actual time=104..104 rows=1 loops=1)
    -> Filter: (table_entity.deleted = false)  (cost=70841 rows=12604) (actual time=0.0496..90.5 rows=246627 loops=1)
        -> Covering index scan on table_entity using table_entity_deleted  (cost=70841 rows=126041) (actual time=0.0484..67.5 rows=246673 loops=1)

-- WITHOUT INDEX
-> Aggregate: count(0)  (cost=74359 rows=1) (actual time=2253..2253 rows=1 loops=1)
    -> Filter: (table_entity.deleted = false)  (cost=73099 rows=12604) (actual time=0.911..2235 rows=246627 loops=1)
        -> Covering index scan on table_entity using index_table_entity_deleted  (cost=73099 rows=126041) (actual time=0.907..2191 rows=246673 loops=1)
```
The actual time goes from `actual time=2253..2253` to `actual time=104..104`.

We have added this index to most of the entities.

### 2. `COUNT(*)` vs. `COUNT(fqnHash)`

Checking on RDS MySQL, we can see how a `COUNT(*)` takes around 40s. Instead, `COUNT(fqnHash)` becomes <1 sec.

```
EXPLAIN ANALYZE select count(*) from table_entity;
-- -> Count rows in table_entity  (actual time=40077..40077 rows=1 loops=1)

EXPLAIN ANALYZE select count(fqnHash) from table_entity;
-- -> Aggregate: count(table_entity.fqnHash)  (cost=75987 rows=1) (actual time=688..688 rows=1 loops=1)
--    -> Covering index scan on table_entity using fqnHash  (cost=63383 rows=126041) (actual time=0.739..674 rows=246673 loops=1)
```

My understanding was that the optimizer would pick up the best route with `COUNT(*)`, but the results are consistently faster with `COUNT(fqnHash)`. We noticed last week with @tutte that the `SELECT COUNT(*) FROM table_entity` in the Sandbox BETA RDS also took a huge toll. Running live the `COUNT(fqnHash)` also gave better results even when the db was under a big load.

Therefore, we have updated the `COUNT` queries to use `getColumnHash`.

Interestingly, with Postgres RDS (~6k tables) the `COUNT(*)` performs better than `COUNT(fqnHash)`:

```
explain analyze select count(fqnhash) from table_entity;
/*
 * Aggregate  (cost=1005.24..1005.25 rows=1 width=8) (actual time=2.603..2.603 rows=1 loops=1)
  ->  Seq Scan on table_entity  (cost=0.00..990.59 rows=5859 width=135) (actual time=0.009..2.228 rows=5859 loops=1)
Planning Time: 0.069 ms
Execution Time: 2.636 ms
 */

explain analyze select count(*) from table_entity;
/*
Aggregate  (cost=1005.24..1005.25 rows=1 width=8) (actual time=1.601..1.601 rows=1 loops=1)
  ->  Seq Scan on table_entity  (cost=0.00..990.59 rows=5859 width=0) (actual time=0.008..1.249 rows=5859 loops=1)
Planning Time: 0.064 ms
Execution Time: 1.627 ms
*/
```

So we're keeping `COUNT(*)` in Postgres.

Note that all these results are maintained even after analyzing the tables.

### 3. Missing Indexes

Adding missing indexes for a couple of tables:

```SQL
ALTER TABLE apps_extension_time_series ADD INDEX apps_extension_time_series_index(appId);
ALTER TABLE suggestions ADD INDEX index_suggestions_type(suggestionType);
ALTER TABLE suggestions ADD INDEX index_suggestions_status(status);
```

### 4. `ANALYZE` in openmetadata-ops

We have added the `analyze-tables` command. It might make sense to have that as another step in the release process to ensure indexes are kept up-to-date.

<!-- For frontend related change, please add screenshots and/or videos of your changes preview! -->

#
### Type of change:
<!-- You should choose 1 option and delete options that aren't relevant -->
- [ ] Bug fix
- [ ] Improvement
- [ ] New feature
- [ ] Breaking change (fix or feature that would cause existing functionality to not work as expected)
- [ ] Documentation

#
### Checklist:
<!-- add an x in [] if done, don't mark items that you didn't do !-->
- [x] I have read the [**CONTRIBUTING**](https://docs.open-metadata.org/developers/contribute) document.
- [ ] My PR title is `Fixes <issue-number>: <short explanation>`
- [ ] I have commented on my code, particularly in hard-to-understand areas. 
- [ ] For JSON Schema changes: I updated the migration scripts or explained why it is not needed.

<!-- Based on the type(s) of your change, uncomment the required checklist 👇 -->

<!-- Bug fix
- [ ] I have added a test that covers the exact scenario we are fixing. For complex issues, comment the issue number in the test for future reference.
-->

<!-- Improvement
- [ ] I have added tests around the new logic.
- [ ] For connector/ingestion changes: I updated the documentation.
-->

<!-- New feature
- [ ] The issue properly describes why the new feature is needed, what's the goal, and how we are building it. Any discussion
    or decision-making process is reflected in the issue.
- [ ] I have updated the documentation.
- [ ] I have added tests around the new logic.
-->

<!-- Breaking change
- [ ] I have added the tag `Backward-Incompatible-Change`.
-->